Version 4.426
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@1957 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ toro/TC_Common/TC_NameIndex.cxx
@@ -184,10 +184,10 @@ void TC_NameIndex_c::ExtractString(
          *psrNameIndex += this->srName_;
          *psrNameIndex += "\"";
 
-	 if( this->index_ != SIZE_MAX )
-	 {
-   	    char szIndex[TIO_FORMAT_STRING_LEN_VALUE];
-            sprintf( szIndex, "%lu", this->index_ );
+         if( this->index_ != SIZE_MAX )
+         {
+            char szIndex[TIO_FORMAT_STRING_LEN_VALUE];
+            sprintf( szIndex, "%lu", static_cast< unsigned long >( this->index_ ));
 
             *psrNameIndex += " ";
             *psrNameIndex += szIndex;
